Combine client and server download pages
