+++ docs/general/clients/index.mdx
@@ -15,7 +15,7 @@ If you are interested in helping out, please see our [contribution guide](/docs/
 
 ## Client List
 
-A list of all Jellyfin clients can be found on the main [Clients page](/clients).
+A list of all Jellyfin clients can be found on the main [Clients page](/downloads/clients).
 
 :::note
