+++ docs/index.md
@@ -19,7 +19,7 @@ title: Introduction
 
 Jellyfin is a Free Software Media System that puts you in control of managing and streaming your media. It is an alternative to the proprietary Emby and Plex, to provide media from a dedicated server to end-user devices via multiple apps. Jellyfin is descended from Emby's 3.5.2 release and ported to the .NET Core framework to enable full cross-platform support. There are no strings attached, no premium licenses or features, and no hidden agendas: just a team who want to build something better and work together to achieve it. We welcome anyone who is interested in joining us in our quest!
 
-You can find a list of all available clients [here](/clients). For more information please see our [about page](/docs/general/about) or the [FAQ](/docs/general/faq). If you are looking for help, check out [this page](/docs/general/getting-help) for all the different communication channels we use.
+You can find a list of all available clients [here](/downloads/clients). For more information please see our [about page](/docs/general/about) or the [FAQ](/docs/general/faq). If you are looking for help, check out [this page](/docs/general/getting-help) for all the different communication channels we use.
 
 Note: Jellyfin is a fast-moving project that is in its early stages, and this documentation as well as the code may change frequently. Please check back often and do not hesitate to contact us via our Matrix channels or the subreddit!
