--- src/pages/clients/index.module.scss
@@ -1,19 +1,0 @@
-.header {
-  display: flex;
-}
-
-@media (max-width: 500px) {
-  .header {
-    flex-direction: column;
-  }
-}
-
-.header .pills {
-  flex-grow: 1;
-}
-
-.pills {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
