+++ src/pages/downloads/docker.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import DownloadsPage from '.';
+import DownloadsPage from './server';
 import { OsType } from '../../data/downloads';
 
 export default function DockerDownloads() {
