+++ src/pages/downloads/linux.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import DownloadsPage from '.';
+import DownloadsPage from './server';
 
 export default function LinuxDownloads() {
   return <DownloadsPage />;
