+++ src/pages/downloads/macos.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import DownloadsPage from '.';
+import DownloadsPage from './server';
 import { OsType } from '../../data/downloads';
 
 export default function MacOSDownloads() {
